+++ Src/Commands/Abstract.cs
@@ -21,6 +21,9 @@ namespace Mdbc.Commands
 {
 	public abstract class Abstract : PSCmdlet
 	{
+		public const string TextParameterQuery = "Parameter Query must be specified and cannot be null.";
+		public const string TextParameterUpdate = "Parameter Update must be specified and cannot be null.";
+		
 		protected virtual void WriteException(Exception exception, object target)
 		{
 			WriteError(new ErrorRecord(exception, "Mdbc", ErrorCategory.NotSpecified, target));